Update HDR Output Control When Enabling Format
If disabling a format that's currently in use, the HDR conversion mode
is set to SYSTEM even if it's already SYSTEM. This is done to make sure
the new set of enabled types is propagated.

Enabling a disabled format was not propagated. Enabling a type that
SYSTEM conversion would prefer would not cause the conversion target to
update. Similarly to propagating disabling a type, fix this by setting
the conversion mode to SYSTEM when it's currently SYSTEM.

Bug: b/319545979
Test: atest PreferredDynamicRangeFragmentTest, manual test on device
Change-Id: I2a7186c821a581b8701ec53b92d04e4ea5db75cf