Fix accessibility issue with "Time to take action" page.
Talkback considers `PreferenceCategory` as a row even when it's title is not set. As a result, it announces `Default` preference as "Row 2` and not `Row 1`. While we couldn't figure out a solution to fix this problem, we realized that usage of `PreferenceCategory` itself is redundant here as there is only one group of related preferences. This, in turn, fixes the original problem.

Test: Manually with udc-tv-dev branch.

Bug: 323997592

Change-Id: Ib4de0ed99ffabe3adcdefd95dfde4e5d5af71076