Fix accessiblity issue with "Text scaling" page.
Talkback considers `PreferenceCategory` as a row even when it's title is not set. As a result, it announces `Small` preference as `Row 2` and not `Row 1`. While we couldn't figure out a solution to fix this problem, we realized that usage of `PreferenceCategory` itself is redundant here as there is only one group of related preferences. This, in turn, fixes the original problem.

Test: Manually with udc-tv-dev branch.
Change-Id: If0286bf6bab44453b7c4cbfe57f66678b00a428f