Copy needed portions of framework and support slice code into TVSettings
Both android.app.slice and androidx.slice are deprecated and TVSettings
only uses a small portion of slices framework as a data format for 1p/3p
customizable settings. This copies needed code into TwoPanelSettingsLib
so that we can support and enhance it for our own use, including support
for updatable settings tee.

Code is copied as is without structural cleanup and refactoring. Individual
parts will be refactored in follow up changes to make specific enhancements.

Bug: 367439361
Flag: NONE not needed

Change-Id: I8ef1d908b01f256a767e43a505a1c416960438e9