Fix a crash in btservices library caused by unused twopanelsettings code
We are moving to a private copy of slices support library that can be extended to our needs, but existing slice providers will not be all migrated at the same time. Fixing a crash in btservices caused by unnecessarily accessing a static variable which is not set because btservices still uses androidx slices.

Bug: 367439361
Flag: NONE not needed
Change-Id: I5d8f0cc4ee344e8d2c9507c38be4521643774d9b