Convert slice support code and main fragment to kotlin
Updatable settings are built using kotlin facilities such as coroutines and it's easier to use them from kotlin code with lifecycle scopes than to convert to futures and worry about timely cancellation. This converts classes that are likely to be modified to kotlin without any other changes, to avoid mixing language migration and behavior changes.

Bug: 367439361
Flag: NONE not needed
Change-Id: I1d4246e3fd23b42cfb60e5d6cb01f750a84309c8