Changing colors of controls to meet contrast ratio requirements.
We change the colors of the switches, radio buttons, and the checkboxes with color state lists. For the switches, we change the Switch to MaterialSwitch by changing the widgetLayout.

We need to handle TVInputMethodPreference in a special way because we can have a situation where the row can be in state ENABLED, but we want the switch can be in state DISABLED, so duplicateStateList will not work.

In order to have this work on U and before, we need to backport ag/25168273 first.

Bug:326269726
Flag: NONE Not needed

Change-Id: Ie499a5ada386a801f568001b54ebb5b235cbaa47