Change the forceFocusableViews logic to check spoken accessibility
With the original logic, when any of the accessibility method was
enabled the some text is forced to be focusable. However, for some
accessibility method, focus on the text doesn't make sense (e.g. Switch
Access). It seems we make it focusable to let the screen reader to read
our the text. This change check the FEEDBACK_SPOKEN instead.

Bug: 321896429
Bug: 375566469
Test: build and run with only Screen Reader enable and with only Switch
Access enabled.
Flag: EXEMPT bugfix

Change-Id: I3c1850030fe38c91988d646447ec3dbddae78d98