TwoPanelSettings Background Color Changes
These changes are necessary to support the behavior changes from the spec.

Specifically, instead of solid colored backgrounds, we use transparent panels for the main panel.

This is an issue because: clipChildren is currently set to false for the preference row container (and all views), so if you scroll up, it will show up under the title view, since it no longer has a solid colored background.

We need a different approach: instead we need to stop the settings rows from drawing outside the preference list container.

We do this by changing the clipping behavior. Before, clipChildren was set to false for all views. Instead, we now explicitly decide when clipChildren should be false, usually clipChildren should only be set to false on specific rows.

We also need to change margins to support this change, specifically we need to add padding in places so that we don’t draw outside of views.

Screenshots:
https://drive.google.com/drive/folders/1bcjc5U1V0uCF0vDqk9qyI3qu44e9Ss7f?usp=sharing

Figma mocks:
https://www.figma.com/design/fc0loHb26sLxwyJUHi4Y2R/Settings-2.0?node-id=81-5592&t=AyB7FQwR9H6PDmNZ-1

Flag: NONE not needed
Bug: 355462390


Change-Id: I239c1c344109bc28b9d179eaddeb97efb4bb2fcd